Added code to identify missing values
I did some homework on a few of the features in the data.  There are a few with missing values occupying more than 90% of the data for that particular feature.  Some of these can probably be dropped, but others might be very important to our analysis.  I also figured out what "BIN" was and replaced the missing values for what they should be.